+++ internal/cmd/cmdlipuninstall/cmdlipuninstall.go
@@ -111,7 +111,7 @@ func askForConfirmation(ctx *context.Context,
 			return fmt.Errorf("failed to get installed tooth metadata: %w", err)
 		}
 
-		log.Infof("  %v: %v", toothRepoPath,
+		log.Infof("  %v@%v: %v", toothRepoPath, metadata.Version().String(),
 			metadata.Info().Name)
 	}