+++ CHANGELOG.md
@@ -7,6 +7,13 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [0.18.1] - 2024-01-12
+
+### Fixed
+
+- Platform-specific items in tooth.json.
+- Go module path escaping.
+
 ## [0.18.0] - 2024-01-12
 
 ### Added
@@ -341,7 +348,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - Basic functions: cache, install, list, show, tooth init, and uninstall.
 
-[unreleased]: https://github.com/lippkg/lip/compare/v0.18.0...HEAD
+[unreleased]: https://github.com/lippkg/lip/compare/v0.18.1...HEAD
+[0.18.1]: https://github.com/lippkg/lip/compare/v0.18.0...v0.18.1
 [0.18.0]: https://github.com/lippkg/lip/compare/v0.17.0...v0.18.0
 [0.17.0]: https://github.com/lippkg/lip/compare/v0.16.1...v0.17.0
 [0.16.1]: https://github.com/lippkg/lip/compare/v0.16.0...v0.16.1